requirements.txt
kfp==1.8.13
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2
mlflow==2.7.1
google-cloud-storage==1.42.3
google-auth<2,>=1.6.1
gcsfs==2023.10.0
